Rename pipelines -> analysis_levels
- Rename module to better reflect its use
- Moved `launch` to be part of `analysis_levels` module.
- Updated `__init__.py`'s as necessary.
