+++ src/niftyone/__init__.py
@@ -1,4 +1,3 @@
 """Large-scale neuroimaging visualization using FiftyOne."""
 
 from ._version import __version__, __version_tuple__
-from .launch import launch
