+++ src/niftyone/analysis_levels/__init__.py
@@ -1,0 +1,5 @@
+"""Initialize module for generating pipelines."""
+
+from .group import group
+from .launch import launch
+from .participant import participant
