--- src/niftyone/pipelines/__init__.py
@@ -1,6 +1,0 @@
-"""Initialize module for generating pipelines."""
-
-from .group import group_pipeline as group
-from .participant import (
-    participant_pipeline as participant,
-)
